cc1352p7: Add a way to do a system reset with OpenOCD
Currently, OpenOCD only reset the CPU because a system reset stop the JTAG
and is not correctly managed by the JTAG driver.
There are many situations where CPU reset is not enough
(radio development is one of them).
This workaround adds support system reset. JTAG connection will lost
but this is OK because we will go through a full JTAG connection process
everytime.
OpenOCD will print a lot of error when the JTAG connection is lost but
they can be safely ignored.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>