drivers: led: Provide a fallback implementation of blink API
The LED blink API i squite useful but this is not supported by all
the LED controller.
This provides a software implementation of blink that use a delayable
work to blink a LED.
This could be enabled using Kconfig, as well the maximum number of LED
that could blink at the same time using blink fallback.
This does not requires any changes to driver to be used.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>